+++ libfreeipmi/src/api/ipmi-openipmi-driver-api.h
@@ -31,11 +31,11 @@ int ipmi_openipmi_cmd_api_ipmb (ipmi_ctx_t ctx,
                                 fiid_obj_t obj_cmd_rq,
                                 fiid_obj_t obj_cmd_rs);
 
-int32_t ipmi_openipmi_cmd_raw_api (ipmi_ctx_t ctx,
-                                   uint8_t *buf_rq,
-                                   size_t buf_rq_len,
-                                   uint8_t *buf_rs,
-                                   size_t buf_rs_len);
+int ipmi_openipmi_cmd_raw_api (ipmi_ctx_t ctx,
+                               uint8_t *buf_rq,
+                               unsigned int buf_rq_len,
+                               uint8_t *buf_rs,
+                               unsigned int buf_rs_len);
 
 #endif /* ipmi-openipmi-driver-api.h */
